+++ tests/unit/test_kernel.py
@@ -36,11 +36,16 @@ from scipy.stats import norm as scipy_norm
 from coreax.data import Data
 from coreax.kernel import (
     AdditiveKernel,
+    ExponentialKernel,
     Kernel,
     LaplacianKernel,
     LinearKernel,
+    LocallyPeriodicKernel,
     PCIMQKernel,
+    PeriodicKernel,
+    PolynomialKernel,
     ProductKernel,
+    RationalQuadraticKernel,
     SquaredExponentialKernel,
     SteinKernel,
 )
@@ -99,7 +104,7 @@ class KernelMeanTest(Generic[_Kernel]):
         self,
         jit_variant: Callable[[Callable], Callable],
         kernel: _Kernel,
-        block_size: Union[int, float, None],
+        block_size: Union[int, None, tuple[Union[int, None], Union[int, None]]],
         axis: Union[int, None],
     ) -> None:
         """
@@ -176,12 +181,14 @@ class KernelGradientTest(ABC, Generic[_Kernel]):
 
     @pytest.mark.parametrize("mode", ["grad_x", "grad_y", "divergence_x_grad_y"])
     @pytest.mark.parametrize("elementwise", [False, True])
+    @pytest.mark.parametrize("auto_diff", [False, True])
     def test_gradients(
         self,
         gradient_problem: tuple[Array, Array],
         kernel: _Kernel,
         mode: Literal["grad_x", "grad_y", "divergence_x_grad_y"],
         elementwise: bool,
+        auto_diff: bool,
     ):
         """Test computation of the kernel gradients."""
         x, y = gradient_problem
@@ -193,7 +200,16 @@ class KernelGradientTest(ABC, Generic[_Kernel]):
         expected_output = getattr(self, reference_mode)(x, y, kernel)
         if elementwise:
             expected_output = expected_output.squeeze()
-        output = getattr(kernel, test_mode)(x, y)
+        if auto_diff:
+            if isinstance(kernel, (AdditiveKernel, ProductKernel)):
+                pytest.skip(
+                    "Autodiff of Additive and Product kernels is tested implicitly."
+                )
+            # Access overridden parent methods that use auto-differentiation
+            autodiff_kernel = super(type(kernel), kernel)
+            output = getattr(autodiff_kernel, test_mode)(x, y)
+        else:
+            output = getattr(kernel, test_mode)(x, y)
         np.testing.assert_array_almost_equal(output, expected_output, decimal=3)
 
     @abstractmethod
@@ -213,6 +229,63 @@ class KernelGradientTest(ABC, Generic[_Kernel]):
         self, x: ArrayLike, y: ArrayLike, kernel: _Kernel
     ) -> Union[Array, np.ndarray]:
         """Compute expected divergence of the kernel w.r.t ``x`` gradient ``y``."""
+
+
+class TestKernelMagicMethods:
+    """Test that the Kernel magic methods produce correct instances."""
+
+    @pytest.mark.parametrize(
+        "mode",
+        [
+            "add_int",
+            "add_float",
+            "add_self",
+            "right_add",
+            "mul_int",
+            "mul_float",
+            "mul_self",
+            "right_mul",
+            "int_pow",
+            "float_pow",
+            "invalid_pow",
+            "invalid_paired_kernel_inputs",
+        ],
+    )
+    def test_magic_methods(  # noqa: C901
+        self, mode: str
+    ):
+        """Test kernel magic methods produce correct paired Kernels."""
+        kernel = LinearKernel()
+        if mode == "add_int":
+            assert kernel + 1 == AdditiveKernel(kernel, LinearKernel(0, 1))
+        elif mode == "add_float":
+            assert kernel + 1.0 == AdditiveKernel(kernel, LinearKernel(0, 1.0))
+        elif mode == "add_self":
+            assert kernel + kernel == AdditiveKernel(kernel, kernel)
+        elif mode == "right_add":
+            assert 1 + kernel == AdditiveKernel(kernel, LinearKernel(0, 1.0))
+        elif mode == "mul_int":
+            assert kernel * 1 == ProductKernel(kernel, LinearKernel(0, 1))
+        elif mode == "mul_float":
+            assert kernel * 1.0 == ProductKernel(kernel, LinearKernel(0, 1.0))
+        elif mode == "mul_self":
+            assert kernel * kernel == ProductKernel(kernel, kernel)
+        elif mode == "right_mul":
+            assert 1 * kernel == ProductKernel(kernel, LinearKernel(0, 1.0))
+        elif mode == "int_pow":
+            assert kernel**4 == ProductKernel(
+                ProductKernel(kernel, kernel), ProductKernel(kernel, kernel)
+            )
+        elif mode == "float_pow":
+            assert kernel**2.6 == ProductKernel(kernel, kernel)  # pyright: ignore
+        elif mode == "invalid_pow":
+            with pytest.raises(ValueError):
+                # pylint: disable=pointless-statement
+                kernel**0.5  # pyright: ignore
+                # pylint: enable=pointless-statement
+        elif mode == "invalid_paired_kernel_inputs":
+            with pytest.raises(ValueError):
+                AdditiveKernel(1, "string")  # pyright: ignore
 
 
 class _MockedPairedKernel:
@@ -391,7 +464,7 @@ class TestProductKernel(
         self, request: pytest.FixtureRequest, kernel: ProductKernel
     ) -> _Problem:
         r"""
-        Test problems for the Product kernel where we multiply two Linear kernels.
+        Test problems for the Product kernel.
 
         Given kernel functions :math:`k:\mathbb{R}^d \times \mathbb{R}^d \to \mathbb{R}`
         and :math:`l:\mathbb{R}^d \times \mathbb{R}^d \to \mathbb{R}`, define the
@@ -484,6 +557,33 @@ class TestProductKernel(
 
         return np.array(expected_divergences)
 
+    def test_symmetric_product_kernel(self):
+        """
+        Test reduced computation capacity of ProductKernel.
+
+        We consider a product kernel with equal input kernels and check that
+        the second kernel is never called.
+        """
+        x = np.array([1])
+
+        # Form two simple mocked kernels and force any == operation to return True
+        first_kernel = MagicMock(spec=Kernel)
+        first_kernel.compute_elementwise.return_value = np.array(1.0)
+        first_kernel.__eq__.return_value = True  # pyright: ignore
+        second_kernel = MagicMock(spec=Kernel)
+
+        # Build Product kernel from mocks and check the second kernel is never called
+        symmetric_kernel = ProductKernel(first_kernel, second_kernel)
+
+        symmetric_kernel.compute_elementwise(x, x)
+        second_kernel.compute_elementwise.assert_not_called()
+        symmetric_kernel.grad_x_elementwise(x, x)
+        second_kernel.compute_elementwise.assert_not_called()
+        symmetric_kernel.grad_y_elementwise(x, x)
+        second_kernel.compute_elementwise.assert_not_called()
+        symmetric_kernel.divergence_x_grad_y_elementwise(x, x)
+        second_kernel.compute_elementwise.assert_not_called()
+
 
 class TestLinearKernel(
     BaseKernelTest[LinearKernel],
@@ -566,6 +666,140 @@ class TestLinearKernel(
             for y_idx in range(y.shape[0]):
                 expected_divergence[x_idx, y_idx] = kernel.output_scale * dimension
         return expected_divergence
+
+
+class TestPolynomialKernel(
+    BaseKernelTest[PolynomialKernel],
+    KernelMeanTest[PolynomialKernel],
+    KernelGradientTest[PolynomialKernel],
+):
+    """Test ``coreax.kernel.PolynomialKernel``."""
+
+    @pytest.fixture(scope="class")
+    def kernel(self) -> PolynomialKernel:
+        random_seed = 2_024
+        parameters = jnp.abs(jr.normal(key=jr.key(random_seed), shape=(3,)))
+        return PolynomialKernel(
+            output_scale=parameters[0].item(),
+            constant=parameters[1].item(),
+            degree=int(jnp.ceil(jnp.abs(parameters[2]))) + 1,
+        )
+
+    @pytest.fixture(params=["floats", "vectors", "arrays"])
+    def problem(self, request, kernel: PolynomialKernel) -> _Problem:
+        r"""
+        Test problems for the Polynomial kernel.
+
+        Given :math:`\rho =`'output_scale', :math:`c =`'constant', and
+        :math:`d=`'degree', the polynomial kernel is defined as
+        :math:`k: \mathbb{R}^d\times \mathbb{R}^d \to \mathbb{R}`,
+        :math:`k(x, y) = \rho (x^Ty + c)^d`.
+
+        We consider the following cases:
+        - `floats`: where x and y are floats
+        - `vectors`: where x and y are vectors of the same size
+        - `arrays`: where x and y are arrays of the same shape
+        """
+        mode = request.param
+        x = 0.5
+        y = 2.0
+        if mode == "floats":
+            expected_distances = 1.0
+        elif mode == "vectors":
+            x = 1.0 * np.arange(3)
+            y = x + 1.0
+            expected_distances = 64.0
+        elif mode == "arrays":
+            x = np.array(([0, 1, 2], [3, 4, 5]))
+            y = np.array(([1, 2, 3], [4, 5, 6]))
+            expected_distances = np.array([[64, 289], [676, 3844]])
+        else:
+            raise ValueError("Invalid problem mode")
+        output_scale = 1.0
+        constant = 0.0
+        degree = 2
+        modified_kernel = eqx.tree_at(
+            lambda x: x.output_scale,
+            eqx.tree_at(
+                lambda x: x.constant,
+                eqx.tree_at(lambda x: x.degree, kernel, degree),
+                constant,
+            ),
+            output_scale,
+        )
+        return _Problem(x, y, expected_distances, modified_kernel)
+
+    def expected_grad_x(
+        self, x: ArrayLike, y: ArrayLike, kernel: PolynomialKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_gradients = np.zeros((num_points, num_points, dimension))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                expected_gradients[x_idx, y_idx] = (
+                    kernel.output_scale
+                    * kernel.degree
+                    * y[y_idx]
+                    * (np.dot(x[x_idx], y[y_idx]) + kernel.constant)
+                    ** (kernel.degree - 1)
+                )
+        return expected_gradients
+
+    def expected_grad_y(
+        self, x: ArrayLike, y: ArrayLike, kernel: PolynomialKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_gradients = np.zeros((num_points, num_points, dimension))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                expected_gradients[x_idx, y_idx] = (
+                    kernel.output_scale
+                    * kernel.degree
+                    * x[x_idx]
+                    * (np.dot(x[x_idx], y[y_idx]) + kernel.constant)
+                    ** (kernel.degree - 1)
+                )
+        return expected_gradients
+
+    def expected_divergence_x_grad_y(
+        self, x: ArrayLike, y: ArrayLike, kernel: PolynomialKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_divergence = np.zeros((num_points, num_points))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                expected_divergence[x_idx, y_idx] = (
+                    kernel.output_scale
+                    * kernel.degree
+                    * (
+                        (
+                            (kernel.degree - 1)
+                            * np.dot(x[x_idx], y[y_idx])
+                            * (np.dot(x[x_idx], y[y_idx]) + kernel.constant)
+                            ** (kernel.degree - 2)
+                        )
+                        + (
+                            dimension
+                            * (np.dot(x[x_idx], y[y_idx]) + kernel.constant)
+                            ** (kernel.degree - 1)
+                        )
+                    )
+                )
+        return expected_divergence
+
+    @pytest.mark.parametrize(
+        "degree", [2.6, 1], ids=["float_degree", "degree_less_than_min"]
+    )
+    def test_invalid_inputs(self, degree):
+        """Test that polynomial kernel rejects bad degree inputs."""
+        with pytest.raises(ValueError):
+            PolynomialKernel(degree=degree)
 
 
 class TestSquaredExponentialKernel(
@@ -721,6 +955,518 @@ class TestSquaredExponentialKernel(
                     * (dimension - dot_product / kernel.length_scale**2)
                 )
         return expected_divergence
+
+
+class TestExponentialKernel(
+    BaseKernelTest[ExponentialKernel],
+    KernelMeanTest[ExponentialKernel],
+    KernelGradientTest[ExponentialKernel],
+):
+    """Test ``coreax.kernel.ExponentialKernel``."""
+
+    @pytest.fixture(scope="class")
+    def kernel(self) -> ExponentialKernel:
+        random_seed = 2_024
+        parameters = jnp.abs(jr.normal(key=jr.key(random_seed), shape=(2,)))
+        return ExponentialKernel(
+            length_scale=parameters[0].item(), output_scale=parameters[1].item()
+        )
+
+    @pytest.fixture(params=["floats", "vectors", "arrays"])
+    def problem(self, request, kernel: ExponentialKernel) -> _Problem:  # noqa: C901
+        r"""
+        Test problems for the Exponential kernel.
+
+        Given :math:`\lambda =`'length_scale' and :math:`\rho =`'output_scale', the
+        exponential kernel is defined as
+        :math:`k: \mathbb{R}^d\times \mathbb{R}^d \to \mathbb{R}`,
+        :math:`k(x, y) = \rho * \exp(-\frac{||x-y||}{2 \lambda^2})` where
+        :math:`||\cdot||` is the usual :math:`L_2`-norm.
+
+        We consider the following cases:
+        1. length scale is :math:`\sqrt{\pi} / 2`:
+        - `floats`: where x and y are floats
+        - `vectors`: where x and y are vectors of the same size
+        - `arrays`: where x and y are arrays of the same shape
+
+        2. length scale or output scale is degenerate:
+        - `negative_length_scale`: should give same result as positive equivalent
+        - `large_negative_length_scale`: should approximately equal one
+        - `near_zero_length_scale`: should approximately equal zero
+        - `negative_output_scale`: should negate the positive equivalent.
+        """
+        length_scale = np.sqrt(np.float32(np.pi) / 2.0)
+        output_scale = 1.0
+        mode = request.param
+        x = 0.5
+        y = 2.0
+        if mode == "floats":
+            expected_distances = 0.6203541
+        elif mode == "vectors":
+            x = 1.0 * np.arange(4)
+            y = x + 1.0
+            expected_distances = 0.5290778
+        elif mode == "arrays":
+            x = np.array(([0, 1, 2, 3], [5, 6, 7, 8]))
+            y = np.array(([1, 2, 3, 4], [5, 6, 7, 8]))
+            expected_distances = np.array([[0.5290778, 0.04145699], [0.07835708, 1.0]])
+        elif mode == "negative_length_scale":
+            length_scale = -length_scale
+            expected_distances = 0.6203541
+        elif mode == "large_negative_length_scale":
+            length_scale = -10000.0
+            expected_distances = 1.0
+        elif mode == "near_zero_length_scale":
+            length_scale = -0.0000001
+            expected_distances = 0.0
+        elif mode == "negative_output_scale":
+            output_scale = -1
+            expected_distances = -0.6203541
+        else:
+            raise ValueError("Invalid problem mode")
+        modified_kernel = eqx.tree_at(
+            lambda x: x.output_scale,
+            eqx.tree_at(lambda x: x.length_scale, kernel, length_scale),
+            output_scale,
+        )
+        return _Problem(x, y, expected_distances, modified_kernel)
+
+    def expected_grad_x(
+        self, x: ArrayLike, y: ArrayLike, kernel: ExponentialKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_gradients = np.zeros((num_points, num_points, dimension))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                sub = x[x_idx, :] - y[y_idx, :]
+                dist = np.linalg.norm(sub)
+                expected_gradients[x_idx, y_idx] = (
+                    -kernel.output_scale
+                    * sub
+                    / (2 * kernel.length_scale**2 * dist)
+                    * np.exp(-dist / (2 * kernel.length_scale**2))
+                )
+
+        return expected_gradients
+
+    def expected_grad_y(
+        self, x: ArrayLike, y: ArrayLike, kernel: ExponentialKernel
+    ) -> np.ndarray:
+        return -self.expected_grad_x(x, y, kernel)
+
+    def expected_divergence_x_grad_y(
+        self, x: ArrayLike, y: ArrayLike, kernel: ExponentialKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_divergence = np.zeros((num_points, num_points))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                sub = np.subtract(x[x_idx], y[y_idx])
+                dist = np.linalg.norm(sub)
+                factor = 2 * kernel.length_scale**2
+                exp = np.exp(-dist / factor)
+
+                first_term = (-exp * sub / dist**2) * ((1 / dist) + 1 / factor)
+                second_term = exp / dist
+
+                expected_divergence[x_idx, y_idx] = (kernel.output_scale / factor) * (
+                    np.dot(first_term, sub) + dimension * second_term
+                )
+        return expected_divergence
+
+
+class TestRationalQuadraticKernel(
+    BaseKernelTest[RationalQuadraticKernel],
+    KernelMeanTest[RationalQuadraticKernel],
+    KernelGradientTest[RationalQuadraticKernel],
+):
+    """Test ``coreax.kernel.RationalQuadraticKernel``."""
+
+    @pytest.fixture(scope="class")
+    def kernel(self) -> RationalQuadraticKernel:
+        random_seed = 2_024
+        parameters = jnp.abs(jr.normal(key=jr.key(random_seed), shape=(2,)))
+        return RationalQuadraticKernel(
+            length_scale=parameters[0].item(), output_scale=parameters[1].item()
+        )
+
+    @pytest.fixture(
+        params=[
+            "floats",
+            "vectors",
+            "arrays",
+            "negative_length_scale",
+            "large_negative_length_scale",
+            "near_zero_length_scale",
+            "negative_output_scale",
+        ]
+    )
+    def problem(  # noqa: C901
+        self, request, kernel: RationalQuadraticKernel
+    ) -> _Problem:
+        r"""
+        Test problems for the RationalQuadratic kernel.
+
+        Given :math:`\lambda =`'length_scale',  :math:`\rho =`'output_scale', and
+        :math:`\alpha =`'relative_weighting', the rational
+        quadratic kernel is defined as
+        :math:`k: \mathbb{R}^d\times \mathbb{R}^d \to \mathbb{R}`,
+        :math:`k(x, y) = \rho * (1 + \frac{||x-y||^2}{2 \alpha \lambda^2})^{-\alpha}`
+        where :math:`||\cdot||` is the usual :math:`L_2`-norm..
+
+        We consider the following cases:
+        1. length scale is :math:`\sqrt{\pi} / 2`:
+        - `floats`: where x and y are floats
+        - `vectors`: where x and y are vectors of the same size
+        - `arrays`: where x and y are arrays of the same shape
+
+        2. length scale, output scale or relative weighting is degenerate:
+        - `negative_length_scale`: should give same result as positive equivalent
+        - `large_negative_length_scale`: should approximately equal one
+        - `near_zero_length_scale`: should approximately equal zero
+        - `negative_output_scale`: should negate the positive equivalent.
+        """
+        length_scale = np.sqrt(np.float32(np.pi) / 2.0)
+        output_scale = 1.0
+        relative_weighting = 1.0
+        mode = request.param
+        x = 0.5
+        y = 2.0
+        if mode == "floats":
+            expected_distances = 0.5826836
+        elif mode == "vectors":
+            x = 1.0 * np.arange(4)
+            y = x + 1.0
+            expected_distances = 0.43990085
+        elif mode == "arrays":
+            x = np.array(([0, 1, 2, 3], [5, 6, 7, 8]))
+            y = np.array(([1, 2, 3, 4], [5, 6, 7, 8]))
+            expected_distances = np.array([[0.43990085, 0.03045903], [0.04679056, 1.0]])
+        elif mode == "negative_length_scale":
+            length_scale = -length_scale
+            expected_distances = 0.5826836
+        elif mode == "large_negative_length_scale":
+            length_scale = -10000.0
+            expected_distances = 1.0
+        elif mode == "near_zero_length_scale":
+            length_scale = -0.0000001
+            expected_distances = 0.0
+        elif mode == "negative_output_scale":
+            output_scale = -1
+            expected_distances = -0.5826836
+        else:
+            raise ValueError("Invalid problem mode")
+        modified_kernel = eqx.tree_at(
+            lambda x: x.output_scale,
+            eqx.tree_at(
+                lambda x: x.length_scale,
+                eqx.tree_at(lambda x: x.relative_weighting, kernel, relative_weighting),
+                length_scale,
+            ),
+            output_scale,
+        )
+        return _Problem(x, y, expected_distances, modified_kernel)
+
+    def expected_grad_x(
+        self, x: ArrayLike, y: ArrayLike, kernel: RationalQuadraticKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_gradients = np.zeros((num_points, num_points, dimension))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                sub = np.subtract(x[x_idx], y[y_idx])
+                expected_gradients[x_idx, y_idx] = -(
+                    kernel.output_scale * sub / kernel.length_scale**2
+                ) * (
+                    1
+                    + np.dot(sub, sub)
+                    / (2 * kernel.relative_weighting * kernel.length_scale**2)
+                ) ** (-kernel.relative_weighting - 1)
+        return expected_gradients
+
+    def expected_grad_y(
+        self, x: ArrayLike, y: ArrayLike, kernel: RationalQuadraticKernel
+    ) -> np.ndarray:
+        return -self.expected_grad_x(x, y, kernel)
+
+    def expected_divergence_x_grad_y(
+        self, x: ArrayLike, y: ArrayLike, kernel: RationalQuadraticKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_divergence = np.zeros((num_points, num_points))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                sub = np.subtract(x[x_idx], y[y_idx])
+                sq_dist = np.dot(sub, sub)
+                div = kernel.relative_weighting * kernel.length_scale**2
+                power = kernel.relative_weighting + 1
+                body = 1 + sq_dist / (2 * div)
+                factor = kernel.output_scale / kernel.length_scale**2
+
+                first_term = factor * body**-power
+                second_term = -(factor * power * sq_dist / div) * body ** -(power + 1)
+
+                expected_divergence[x_idx, y_idx] = dimension * first_term + second_term
+        return expected_divergence
+
+
+class TestPeriodicKernel(
+    BaseKernelTest[PeriodicKernel],
+    KernelMeanTest[PeriodicKernel],
+    KernelGradientTest[PeriodicKernel],
+):
+    """Test ``coreax.kernel.PeriodicKernel``."""
+
+    @pytest.fixture(scope="class")
+    def kernel(self) -> PeriodicKernel:
+        random_seed = 2_024
+        parameters = jnp.abs(jr.normal(key=jr.key(random_seed), shape=(3,)))
+        return PeriodicKernel(
+            length_scale=parameters[0].item(),
+            output_scale=parameters[1].item(),
+            periodicity=parameters[2].item(),
+        )
+
+    @pytest.fixture(params=["floats", "vectors", "arrays"])
+    def problem(self, request, kernel: PeriodicKernel) -> _Problem:  # noqa: C901
+        r"""
+        Test problems for the PeriodicKernel kernel.
+
+        Given :math:`\lambda =`'length_scale',  :math:`\rho =`'output_scale', and
+        :math:`\p =`'periodicity', the periodic kernel is defined as
+        :math:`k: \mathbb{R}^d\times \mathbb{R}^d \to \mathbb{R}`,
+        :math:`k(x, y) = \rho * \exp(\frac{-2 \sin^2(\pi ||x-y||/p)}{\lambda^2})` where
+        :math:`||\cdot||` is the usual :math:`L_2`-norm.
+
+        We consider the following cases:
+        1. length scale is :math:`\sqrt{\pi} / 2`:
+        - `floats`: where x and y are floats
+        - `vectors`: where x and y are vectors of the same size
+        - `arrays`: where x and y are arrays of the same shape
+
+        2. length scale, output scale or periodicity is degenerate:
+        - `negative_length_scale`: should give same result as positive equivalent
+        - `large_negative_length_scale`: should approximately equal one
+        - `near_zero_length_scale`: should approximately equal zero
+        - `negative_output_scale`: should negate the positive equivalent.
+        """
+        length_scale = np.sqrt(np.float32(np.pi) / 2.0)
+        output_scale = 1.0
+        periodicity = 1.0
+        mode = request.param
+        x = 0.5
+        y = 2.0
+        if mode == "floats":
+            expected_distances = 0.27992335
+        elif mode == "vectors":
+            x = 1.0 * np.arange(4)
+            y = 1.0 * np.ones(4)
+            expected_distances = 0.2889656
+        elif mode == "arrays":
+            x = np.array(([1.5, 1, 1, 1], [5, 6, 7, 8]))
+            y = np.array(([1, 2.5, 3, 4], [2.5, 2, 2, 2]))
+            expected_distances = np.array([[0.95196974, 1.0], [0.5552613, 0.8318974]])
+        elif mode == "negative_length_scale":
+            length_scale = -length_scale
+            expected_distances = 0.27992335
+        elif mode == "large_negative_length_scale":
+            length_scale = -10000.0
+            expected_distances = 1.0
+        elif mode == "near_zero_length_scale":
+            length_scale = -0.0000001
+            expected_distances = 0.0
+        elif mode == "negative_output_scale":
+            output_scale = -1
+            expected_distances = -0.27992335
+        else:
+            raise ValueError("Invalid problem mode")
+        modified_kernel = eqx.tree_at(
+            lambda x: x.output_scale,
+            eqx.tree_at(
+                lambda x: x.length_scale,
+                eqx.tree_at(lambda x: x.periodicity, kernel, periodicity),
+                length_scale,
+            ),
+            output_scale,
+        )
+        return _Problem(x, y, expected_distances, modified_kernel)
+
+    def expected_grad_x(
+        self, x: ArrayLike, y: ArrayLike, kernel: PeriodicKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_gradients = np.zeros((num_points, num_points, dimension))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                sub = np.subtract(x[x_idx], y[y_idx])
+                dist = np.linalg.norm(sub)
+                body = np.pi * dist / kernel.periodicity
+
+                expected_gradients[x_idx, y_idx] = (
+                    (
+                        4
+                        * sub
+                        * kernel.output_scale
+                        * np.pi
+                        / (dist * kernel.periodicity * kernel.length_scale**2)
+                    )
+                    * np.sin(body)
+                    * np.cos(body)
+                    * np.exp(-(2 / kernel.length_scale**2) * np.sin(body) ** 2)
+                )
+        return expected_gradients
+
+    def expected_grad_y(
+        self, x: ArrayLike, y: ArrayLike, kernel: PeriodicKernel
+    ) -> np.ndarray:
+        return -self.expected_grad_x(x, y, kernel)
+
+    def expected_divergence_x_grad_y(
+        self, x: ArrayLike, y: ArrayLike, kernel: PeriodicKernel
+    ) -> np.ndarray:
+        x = np.atleast_2d(x)
+        y = np.atleast_2d(y)
+        num_points, dimension = x.shape
+        expected_divergence = np.zeros((num_points, num_points))
+        for x_idx in range(x.shape[0]):
+            for y_idx in range(y.shape[0]):
+                sub = np.subtract(x[x_idx], y[y_idx])
+                dist = np.linalg.norm(sub)
+                factor = np.pi / kernel.periodicity
+
+                func_1 = 1 / dist
+                func_2 = np.sin(factor * dist)
+                func_3 = np.cos(factor * dist)
+                func_4 = np.exp(
+                    -(2 / kernel.length_scale**2) * np.sin(factor * dist) ** 2
+                )
+
+                first_term = func_1 * func_2 * func_3 * func_4
+                second_term = (
+                    -sub / dist * func_1**2 * func_2 * func_3 * func_4
+                    - sub / dist * factor * func_1 * func_2**2 * func_4
+                    + sub / dist * factor * func_1 * func_3**2 * func_4
+                    - (
+                        4
+                        * factor
+                        * kernel.output_scale
+                        / kernel.length_scale**2
+                        * sub
+                        * func_1**2
+                        * func_2**2
+                        * func_3**2
+                        * func_4
+                    )
+                )
+                expected_divergence[x_idx, y_idx] = (
+                    4
+                    * factor
+                    * kernel.output_scale
+                    / kernel.length_scale**2
+                    * (dimension * first_term + np.dot(second_term, sub))
+                )
+        return expected_divergence
+
+
+class TestLocallyPeriodicKernel(
+    BaseKernelTest[LocallyPeriodicKernel],
+    KernelMeanTest[LocallyPeriodicKernel],
+):
+    """Test ``coreax.kernel.LocallyPeriodicKernel``."""
+
+    @pytest.fixture(scope="class")
+    def kernel(self) -> LocallyPeriodicKernel:
+        random_seed = 2_024
+        parameters = jnp.abs(jr.normal(key=jr.key(random_seed), shape=(5,)))
+        return LocallyPeriodicKernel(
+            periodic_length_scale=parameters[0].item(),
+            periodic_output_scale=parameters[1].item(),
+            periodicity=parameters[2].item(),
+            squared_exponential_length_scale=parameters[3].item(),
+            squared_exponential_output_scale=parameters[4].item(),
+        )
+
+    @pytest.fixture(params=["floats", "vectors", "arrays"])
+    def problem(self, request, kernel: LocallyPeriodicKernel) -> _Problem:  # noqa: C901
+        r"""
+        Test problems for the LocallyPeriodicKernel kernel.
+
+        Given :math:`\lambda =`'length_scale',  :math:`\rho =`'output_scale', and
+        :math:`\p =`'periodicity', the periodic kernel is defined as
+        :math:`k: \mathbb{R}^d\times \mathbb{R}^d \to \mathbb{R}`,
+        :math:`k(x, y) = r(x,y)l(x,y)` where :math:`r` is the periodic kernel and
+        :math:`l` is the squared exponential kernel.
+
+        We consider the following cases:
+        1. length scale is :math:`\sqrt{\pi} / 2`:
+        - `floats`: where x and y are floats
+        - `vectors`: where x and y are vectors of the same size
+        - `arrays`: where x and y are arrays of the same shape
+
+        2. length scale, output scale or periodicity is degenerate:
+        - `negative_length_scale`: should give same result as positive equivalent
+        - `large_negative_length_scale`: should approximately equal one
+        - `near_zero_length_scale`: should approximately equal zero
+        - `negative_output_scale`: should negate the positive equivalent.
+        """
+        length_scale = np.sqrt(np.float32(np.pi) / 2.0)
+        output_scale = 1.0
+        periodicity = 1.0
+        mode = request.param
+        x = 0.5
+        y = 2.0
+        if mode == "floats":
+            expected_distances = 0.13677244
+        elif mode == "vectors":
+            x = 1.0 * np.arange(4)
+            y = 1.0 * np.ones(4)
+            expected_distances = 0.04279616
+        elif mode == "arrays":
+            x = np.array(([1.5, 1, 1, 1], [5, 6, 7, 8]))
+            y = np.array(([1, 2.5, 3, 4], [2.5, 2, 2, 2]))
+            expected_distances = np.array(
+                [[6.8532992e-03, 2.7992335e-01], [2.6033020e-09, 2.5797108e-12]]
+            )
+        elif mode == "negative_length_scale":
+            length_scale = -length_scale
+            expected_distances = 0.13677244
+        elif mode == "large_negative_length_scale":
+            length_scale = -10000.0
+            expected_distances = 1.0
+        elif mode == "near_zero_length_scale":
+            length_scale = -0.0000001
+            expected_distances = 0.0
+        elif mode == "negative_output_scale":
+            output_scale = -1
+            expected_distances = -0.13677244
+        else:
+            raise ValueError("Invalid problem mode")
+        modified_kernel = eqx.tree_at(
+            lambda x: x.first_kernel,
+            eqx.tree_at(
+                lambda x: x.second_kernel,
+                kernel,
+                SquaredExponentialKernel(
+                    length_scale=length_scale, output_scale=output_scale
+                ),
+            ),
+            PeriodicKernel(
+                length_scale=length_scale,
+                output_scale=output_scale,
+                periodicity=periodicity,
+            ),
+        )
+        return _Problem(x, y, expected_distances, modified_kernel)
 
 
 class TestLaplacianKernel(